Use the extra fields of the federation protocol.
Turns out it is supported to have all our data entered in Ripple itself.

This means we can largely get rid of our attempts of trying to flexibly
parse a destination string, and just present a proper form to the user.

Also, this finally fixes the currency listing of the federation view.